refactor: remove legacy plotter and expose utilities
