--- He3_Plotter.py
@@ -1,669 +1,0 @@
-import matplotlib
-matplotlib.use('Agg')  # Use non-interactive backend to prevent GUI blocking
-
-from pathlib import Path
-import re
-import pandas as pd
-import matplotlib.pyplot as plt
-import numpy as np
-from tkinter import Tk
-from tkinter.filedialog import askopenfilename, askopenfilenames, askdirectory
-from datetime import datetime
-import logging
-import atexit
-
-logger = logging.getLogger(__name__)
-
-# Global tag appended to output filenames; configurable via set_filename_tag
-FILENAME_TAG = ""
-
-# Default extension used when saving plot images; configurable via
-# :func:`set_plot_extension`.
-PLOT_EXTENSION = "pdf"
-
-# ---- Utility Functions ----
-_hidden_root = None
-
-
-def _get_hidden_root():
-    global _hidden_root
-    if _hidden_root is None:
-        _hidden_root = Tk()
-        _hidden_root.withdraw()
-        atexit.register(_hidden_root.destroy)
-    return _hidden_root
-
-
-def select_file(title="Select a file"):
-    """Prompt the user to select a single file.
-
-    Passing the hidden root as the dialog's parent caused the dialog window to
-    be positioned relative to that (1×1, off‑screen) root on some systems,
-    making it appear partially off screen and immovable.  Instead we just
-    initialise Tk to ensure the Tk subsystem is ready and let the operating
-    system place the dialog normally so it can be moved by the user.
-    """
-    _get_hidden_root()  # Ensure Tk is initialised
-    return askopenfilename(title=title)
-
-
-def select_folder(title="Select a folder"):
-    """Prompt the user to select a folder.
-
-    See :func:`select_file` for an explanation of why the dialog is not given a
-    parent window.
-    """
-    _get_hidden_root()
-    return askdirectory(title=title)
-
-def make_plot_dir(base_path):
-    base_path = Path(base_path)
-    plot_dir = base_path / "plots"
-    plot_dir.mkdir(parents=True, exist_ok=True)
-    return plot_dir
-
-
-def set_filename_tag(tag: str) -> None:
-    """Configure a tag to be appended to output filenames.
-
-    Parameters
-    ----------
-    tag:
-        Arbitrary text entered by the user to help differentiate analyses. It
-        will be inserted into output filenames by :func:`get_output_path`.
-    """
-
-    global FILENAME_TAG
-    FILENAME_TAG = tag.strip()
-
-def set_plot_extension(extension: str) -> None:
-    """Configure the file extension used for saved plot images.
-
-    Parameters
-    ----------
-    extension:
-        Desired file extension (e.g. ``"pdf"``, ``"png"``).  If an empty
-        string is provided, the default ``"pdf"`` is used.
-    """
-
-    global PLOT_EXTENSION
-    ext = extension.strip().lower()
-    PLOT_EXTENSION = ext if ext else "pdf"
-
-# Utility to get output path and ensure directory exists
-def get_output_path(base_path, filename_prefix, descriptor, extension="pdf", subfolder="plots"):
-    """Return a filesystem path for saving output files.
-
-    Any user-provided :data:`FILENAME_TAG` is inserted before the date to help
-    differentiate saved plots and CSVs.
-    """
-
-    base_path = Path(base_path)
-    output_dir = base_path / subfolder
-    output_dir.mkdir(parents=True, exist_ok=True)
-    date_str = datetime.now().strftime("%Y-%m-%d")
-    tag = f" {FILENAME_TAG.strip()}" if FILENAME_TAG.strip() else ""
-    filename = f"{filename_prefix} {descriptor}{tag} {date_str}.{extension}"
-    return output_dir / filename
-
-
-def common_path(paths):
-    paths = [Path(p).resolve() for p in paths]
-    min_len = min(len(p.parts) for p in paths)
-    common_parts = []
-    for i in range(min_len):
-        part = paths[0].parts[i]
-        if all(p.parts[i] == part for p in paths[1:]):
-            common_parts.append(part)
-        else:
-            break
-    return Path(*common_parts)
-
-def process_simulation_file(file_path, area, volume, neutron_yield):
-    df_neutron, _ = read_tally_blocks_to_df(file_path)
-    if df_neutron.empty:
-        return None
-    return calculate_rates(df_neutron, area, volume, neutron_yield)
-
-# ---- Function to read MCNP tally blocks and convert to DataFrame ----
-def read_tally_blocks_to_df(file_path, tally_ids=("14", "24", "34")):
-    with open(file_path, "r", encoding="utf-8", errors="ignore") as f:
-        lines = f.readlines()
-
-    dataframes = {}
-
-    for tally_id in tally_ids:
-        start_index = None
-        for i, line in enumerate(lines):
-            if re.search(rf"1tally\s+{tally_id}\b", line):
-                start_index = i
-                break
-        if start_index is not None:
-            for j in range(start_index, len(lines)):
-                if "energy" in lines[j].lower():
-                    data_lines = lines[j+1:]
-                    break
-            parsed = []
-            for line in data_lines:
-                if not line.strip() or "total" in line.lower():
-                    break
-                parts = line.strip().split()
-                if len(parts) == 3:
-                    try:
-                        energy, value, error = map(float, parts)
-                        parsed.append((energy, value, error))
-                    except:
-                        continue
-            df = pd.DataFrame(parsed, columns=["energy", "value", "error"])
-            dataframes[tally_id] = df
-
-    if "14" not in dataframes or "24" not in dataframes:
-        # If required tallies are missing, return empty DataFrames to avoid
-        # unpacking errors in calling code.
-        logger.warning(f"No valid tally data found in {file_path}")
-        return pd.DataFrame(), pd.DataFrame()
-
-    df_combined = pd.merge(dataframes["14"], dataframes["24"], on="energy", suffixes=("_incident", "_detected"))
-    df_combined.rename(columns={
-        "value_incident": "neutrons_incident_cm2",
-        "error_incident": "frac_error_incident_cm2",
-        "value_detected": "neutrons_detected_cm2",
-        "error_detected": "frac_error_detected_cm2"
-    }, inplace=True)
-    df_photon = dataframes["34"].rename(columns={
-        "energy": "photon_energy",
-        "value": "photons",
-        "error": "photon_error"
-    }) if "34" in dataframes else pd.DataFrame()
-    return df_combined, df_photon
-
-# ---- Function to calculate rates and propagated errors ----
-def calculate_rates(df, area, volume, neutron_yield):
-    df["rate_incident"] = df["neutrons_incident_cm2"] * neutron_yield * area
-    df["rate_detected"] = df["neutrons_detected_cm2"] * neutron_yield * volume
-    df["efficiency"] = df["rate_detected"] / df["rate_incident"]
-
-    df["rate_incident_err2"] = (df["frac_error_incident_cm2"] * df["rate_incident"]) ** 2
-    df["rate_detected_err2"] = (df["frac_error_detected_cm2"] * df["rate_detected"]) ** 2
-    df["rate_incident_err"] = np.sqrt(df["rate_incident_err2"])
-    df["rate_detected_err"] = np.sqrt(df["rate_detected_err2"])
-
-    df["efficiency_err"] = df["efficiency"] * np.sqrt(
-        (df["rate_detected_err"] / df["rate_detected"]) ** 2 +
-        (df["rate_incident_err"] / df["rate_incident"]) ** 2
-    )
-    return df
-
-# ---- Function to plot and save neutron rate and efficiency curves ----
-def plot_efficiency_and_rates(df, filename):
-    filename = Path(filename)
-    base_name = filename.stem
-    base_dir = filename.parent
-
-    plt.figure(figsize=(10, 6))
-    plt.errorbar(df["energy"], df["rate_incident"], yerr=df["rate_incident_err"], label="Incident Rate", fmt='o-', markersize=3, capsize=2)
-    plt.errorbar(df["energy"], df["rate_detected"], yerr=df["rate_detected_err"], label="Detected Rate", fmt='s-', markersize=3, capsize=2)
-    plt.xlabel("Energy (MeV)")
-    plt.ylabel("Neutron Rate")
-    plt.title("Neutron Rates vs Energy")
-    plt.legend()
-    plt.grid(True)
-    plt.semilogx()
-    plt.tight_layout()
-    rate_path = get_output_path(
-        base_dir, base_name, "Neutron rate plot", extension=PLOT_EXTENSION, subfolder="plots"
-    )
-    plt.savefig(rate_path)
-    plt.close()
-    logger.info(f"Saved: {rate_path}")
-
-    plt.figure(figsize=(8, 6))
-    plt.errorbar(df["energy"], df["efficiency"], yerr=df["efficiency_err"], fmt='^-', color='green', markersize=3, capsize=2)
-    plt.xlabel("Energy (MeV)")
-    plt.ylabel("Detection Efficiency")
-    plt.title("Efficiency vs Energy")
-    plt.grid(True)
-    plt.semilogx()
-    plt.tight_layout()
-    eff_path = get_output_path(
-        base_dir, base_name, "efficiency curve", extension=PLOT_EXTENSION, subfolder="plots"
-    )
-    plt.savefig(eff_path)
-    plt.close()
-    logger.info(f"Saved: {eff_path}")
-    # plt.show()
-
-
-# ---- Function to export neutron analysis summary to CSV ----
-def export_summary_to_csv(df, filename):
-    filename = Path(filename)
-    base_name = filename.stem
-    base_dir = filename.parent
-    summary_data = {
-        "Total Incident Neutron": [df["rate_incident"].sum()],
-        "Incident Error": [np.sqrt(df["rate_incident_err2"].sum())],
-        "Total Detected Neutron": [df["rate_detected"].sum()],
-        "Detected Error": [np.sqrt(df["rate_detected_err2"].sum())]
-    }
-    summary_df = pd.DataFrame(summary_data)
-    csv_path = get_output_path(base_dir, base_name, "summary", extension="csv", subfolder="csvs")
-    summary_df.to_csv(csv_path, index=False)
-    logger.info(f"Saved: {csv_path}")
-
-# ---- Function to compute chi-squared statistics between observed and expected values ----
-def calculate_chi_squared(obs, exp, obs_err, exp_err):
-    chi2 = np.sum(((obs - exp) ** 2) / (obs_err ** 2 + exp_err ** 2))
-    dof = len(obs)
-    return chi2, dof, chi2 / dof if dof > 0 else np.nan
-
-# ---- Function to extract moderator thickness from filename ----
-def parse_thickness_from_filename(filename):
-    """Return the moderator thickness parsed from the filename.
-
-    Accepts filenames ending in either ``"_10cmo"`` or the shortened
-    ``"_10o"`` format.  The optional ``"cm"`` substring is handled by the
-    regular expression so both conventions are supported.
-    """
-    match = re.search(r'_(\d+)(?:cm)?o$', filename)
-    return int(match.group(1)) if match else None
-
-# ---- Define constants ----
-SINGLE_SOURCE_YIELD = 2.5e6
-THREE_SOURCE_YIELD = (2.5e6) + (1.25e6) + (7.5e6)
-
-LENGTH_CM = 100.0     # Length of the cylinder in cm
-DIAMETER_CM = 5.0     # Diameter of the cylinder in cm
-RADIUS_CM = DIAMETER_CM / 2.0
-AREA = LENGTH_CM * DIAMETER_CM
-VOLUME = np.pi * (RADIUS_CM ** 2) * LENGTH_CM
-
-EXP_RATE = 247.0333333
-EXP_ERR = 0.907438397
-
-
-# ---- Neutron yield selection logic ----
-def select_neutron_yield():
-    choice = input(
-        "Select neutron yield configuration:\n"
-        "1. Single source (2.5e6 n/s)\n"
-        "2. Three sources (weighted sum)\n"
-        "Enter 1 or 2: "
-    ).strip()
-    if choice == "1":
-        return SINGLE_SOURCE_YIELD
-    elif choice == "2":
-        return THREE_SOURCE_YIELD
-    else:
-        logger.warning("Invalid selection. Defaulting to single source (2.5e6 n/s).")
-        return SINGLE_SOURCE_YIELD
-
-# ---- Start user interaction loop ----
-def prompt_for_valid_file(title="Select MCNP Output File"):
-    while True:
-        file_path = select_file(title)
-        if not file_path:
-            logger.warning("No file selected. Please try again.")
-            continue
-        result = read_tally_blocks_to_df(file_path)
-        if result is not None:
-            return file_path, result
-        logger.error("Invalid file selected. No tally data found. Please select another file.")
-
-def run_analysis_type_1(file_path, area, volume, neutron_yield, export_csv=True):
-    file_path = Path(file_path)
-    df = process_simulation_file(file_path, area, volume, neutron_yield)
-    # --- Export neutron and photon tally blocks to CSV ---
-    if export_csv:
-        df_neutron, df_photon = read_tally_blocks_to_df(file_path)
-        if df_neutron is not None and not df_neutron.empty:
-            neutron_csv_path = get_output_path(
-                file_path.parent, file_path.stem,
-                "neutron tallies", extension="csv", subfolder="csvs"
-            )
-            df_neutron.to_csv(neutron_csv_path, index=False)
-            logger.info(f"Saved: {neutron_csv_path}")
-        if df_photon is not None and not df_photon.empty:
-            photon_csv_path = get_output_path(
-                file_path.parent, file_path.stem,
-                "photon tally", extension="csv", subfolder="csvs"
-            )
-            df_photon.to_csv(photon_csv_path, index=False)
-            logger.info(f"Saved: {photon_csv_path}")
-    if df is None:
-        return
-    logger.info(f"Total Incident Neutron: {df['rate_incident'].sum():.3e} ± {np.sqrt(df['rate_incident_err2'].sum()):.3e}")
-    logger.info(f"Total Detected Neutron: {df['rate_detected'].sum():.3e} ± {np.sqrt(df['rate_detected_err2'].sum()):.3e}")
-    plot_efficiency_and_rates(df, file_path)
-    if export_csv:
-        export_summary_to_csv(df, file_path)
-
-
-
-def run_analysis_type_2(
-    folder_paths,
-    labels=None,
-    lab_data_path=None,
-    area=AREA,
-    volume=VOLUME,
-    neutron_yield=SINGLE_SOURCE_YIELD,
-    export_csv=True,
-):
-    """Plot thickness comparisons from one or more simulation folders.
-
-    Parameters
-    ----------
-    folder_paths : str | list[str]
-        Path(s) to folders containing MCNP output files for each dataset.
-    labels : list[str] | None
-        Optional labels for the datasets. If ``None`` the folder names are used.
-    lab_data_path : str | None
-        Optional path to a CSV file with experimental data containing
-        ``thickness``, ``cps`` and ``error_cps`` columns.
-    """
-
-    if isinstance(folder_paths, (str, Path)):
-        folder_paths = [folder_paths]
-    folder_paths = [Path(p) for p in folder_paths]
-
-    if labels is None:
-        labels = [p.name for p in folder_paths]
-
-    experimental_df = None
-    if lab_data_path:
-        experimental_df = pd.read_csv(lab_data_path)
-        experimental_df.columns = experimental_df.columns.str.strip()
-
-    all_results = []
-    for folder_path, label in zip(folder_paths, labels):
-        results = []
-        for entry in folder_path.iterdir():
-            if not entry.name.endswith('o') or not entry.is_file():
-                continue
-            thickness = parse_thickness_from_filename(entry.name)
-            if thickness is None:
-                continue
-            result = read_tally_blocks_to_df(entry)
-            if result is None:
-                continue
-            df_neutron, _ = result
-            if df_neutron.empty:
-                continue
-            df = calculate_rates(df_neutron, area, volume, neutron_yield)
-            total_detected = df['rate_detected'].sum()
-            total_error = np.sqrt(df['rate_detected_err2'].sum())
-            results.append({
-                'thickness': thickness,
-                'simulated_detected': total_detected,
-                'simulated_error': total_error,
-                'dataset': label,
-            })
-        if results:
-            all_results.append(pd.DataFrame(results).sort_values(by='thickness'))
-
-    if not all_results:
-        logger.warning("No matching simulated CSV files found in folders.")
-        return
-
-    combined_df = pd.concat(all_results, ignore_index=True)
-
-    if export_csv:
-        base_dir = common_path(folder_paths)
-        csv_path = get_output_path(
-            base_dir, 'multi_thickness', 'comparison data', extension='csv', subfolder='csvs'
-        )
-        combined_df.to_csv(csv_path, index=False)
-        logger.info(f"Saved: {csv_path}")
-
-    if experimental_df is not None:
-        for label in combined_df['dataset'].unique():
-            df_label = combined_df[combined_df['dataset'] == label]
-            merged = pd.merge(df_label, experimental_df, on='thickness')
-            if merged.empty:
-                continue
-            chi_squared, dof, reduced_chi_squared = calculate_chi_squared(
-                merged['simulated_detected'],
-                merged['cps'],
-                merged['simulated_error'],
-                merged['error_cps'],
-            )
-            logger.info(
-                f"{label}: Chi-squared: {chi_squared:.2f}, DoF: {dof}, Reduced Chi-squared: {reduced_chi_squared:.2f}"
-            )
-
-    plt.figure(figsize=(10, 6))
-    markers = ['o', 's', '^', 'd', 'v', '<', '>', 'p', 'h', '*']
-    for i, label in enumerate(combined_df['dataset'].unique()):
-        df_label = combined_df[combined_df['dataset'] == label]
-        plt.errorbar(
-            df_label['thickness'],
-            df_label['simulated_detected'],
-            yerr=df_label['simulated_error'],
-            fmt=markers[i % len(markers)],
-            linestyle='-',
-            label=label,
-            capsize=5,
-        )
-    if experimental_df is not None:
-        plt.errorbar(
-            experimental_df['thickness'],
-            experimental_df['cps'],
-            yerr=experimental_df['error_cps'],
-            fmt='k--',
-            label='Experimental',
-            capsize=5,
-        )
-        plt.plot(
-            experimental_df['thickness'],
-            experimental_df['cps'],
-            linestyle='--',
-            color='black',
-        )
-        plt.title('Simulated vs Experimental Neutron Detection')
-    else:
-        plt.title('Simulated Neutron Detection')
-    plt.xlabel('Moderator Thickness (cm)')
-    plt.ylabel('Counts Per Second (CPS)')
-    plt.grid(True)
-    plt.legend()
-    plt.ylim(bottom=0)
-    plt.tight_layout()
-
-    base_dir = common_path(folder_paths)
-    save_path = get_output_path(
-        base_dir, 'multi_thickness', 'comparison plot', extension=PLOT_EXTENSION, subfolder='plots'
-    )
-    plt.savefig(save_path)
-    plt.close()
-    logger.info(f"Saved: {save_path}")
-
-def run_analysis_type_3(folder_path, area, volume, neutron_yield, export_csv=True):
-    folder_path = Path(folder_path)
-    exp_rate = EXP_RATE
-    exp_err = EXP_ERR
-    results = []
-    for entry in folder_path.iterdir():
-        # Only process MCNP output files, which end with "o"
-        if not entry.name.endswith("o") or not entry.is_file():
-            continue
-        match = re.search(r'([-+]?\d+_\d+|\d+)', entry.name)
-        if match:
-            distance_str = match.group(1).replace('_', '.')
-            try:
-                distance = float(distance_str)
-            except ValueError:
-                continue
-            result = read_tally_blocks_to_df(entry)
-            if result is None:
-                continue
-            df_neutron, _ = result
-            df = calculate_rates(df_neutron, AREA, VOLUME, neutron_yield)
-            total_detected = df["rate_detected"].sum()
-            total_error = np.sqrt(df["rate_detected_err2"].sum())
-            results.append({
-                "distance": distance,
-                "rate_detected": total_detected,
-                "rate_error": total_error
-            })
-    if not results:
-        logger.warning("No matching simulated CSV files found in folder.")
-        return
-    distance_df = pd.DataFrame(results).sort_values(by="distance")
-    # Export displacement data to CSV
-    folder_name = folder_path.name
-    if export_csv:
-        csv_path = get_output_path(folder_path, folder_name, "source shift data", extension="csv", subfolder="csvs")
-        distance_df.to_csv(csv_path, index=False)
-        logger.info(f"Saved: {csv_path}")
-    plt.figure(figsize=(10, 6))
-    plt.errorbar(distance_df["distance"], distance_df["rate_detected"], yerr=distance_df["rate_error"], fmt='o', capsize=5, label="Simulated")
-    fit_coeffs, cov_matrix = np.polyfit(distance_df["distance"], distance_df["rate_detected"], 1, cov=True)
-    slope, intercept = fit_coeffs
-    slope_err, intercept_err = np.sqrt(np.diag(cov_matrix))
-    fitted_values = np.polyval(fit_coeffs, distance_df["distance"])
-    residuals = distance_df["rate_detected"] - fitted_values
-    chi_squared_fit = np.sum((residuals / distance_df["rate_error"]) ** 2)
-    dof_fit = len(distance_df) - 2  # 2 parameters in linear fit
-    reduced_chi_squared_fit = chi_squared_fit / dof_fit
-    logger.info(f"Chi-squared of linear fit: {chi_squared_fit:.2f}")
-    logger.info(f"Degrees of freedom: {dof_fit}")
-    logger.info(f"Reduced Chi-squared: {reduced_chi_squared_fit:.2f}")
-    if slope != 0:
-        x_intersect = (exp_rate - intercept) / slope
-        x_intersect_err = x_intersect * np.sqrt((intercept_err / (exp_rate - intercept)) ** 2 + (slope_err / slope) ** 2)
-        logger.info(f"Intersection of fit line with experimental value at x = {x_intersect:.3f} ± {x_intersect_err:.3f} cm")
-        extended_margin = 1.0
-        min_x = min(distance_df["distance"].min(), x_intersect - extended_margin)
-        max_x = max(distance_df["distance"].max(), x_intersect + extended_margin)
-        fit_x = np.linspace(min_x, max_x, 500)
-        fit_y = np.polyval(fit_coeffs, fit_x)
-        plt.plot(fit_x, fit_y, linestyle='--', color='blue', label="Best Fit Line")
-        plt.annotate(
-            f'Intersection: {x_intersect:.2f} ± {x_intersect_err:.2f} cm',
-            xy=(x_intersect, exp_rate),
-            xytext=(x_intersect + 0.5, exp_rate + 0.05 * exp_rate),
-            arrowprops=dict(arrowstyle='->', color='black'),
-            fontsize=10, color='black'
-        )
-    else:
-        logger.warning("Fit line is horizontal; no intersection with experimental line.")
-    plt.axhline(y=exp_rate, color='red', linestyle='--', label=f"Experimental = {exp_rate:.2e}")
-    plt.axhspan(exp_rate - exp_err, exp_rate + exp_err, color='red', alpha=0.2, label="Experimental Uncertainty")
-    plt.xlabel("Source Displacement (cm)")
-    plt.ylabel("Total Detected Rate")
-    plt.title("Detected Rate vs Source Displacement")
-    plt.grid(True)
-    plt.legend()
-    plt.tight_layout()
-    plt.text(
-        0.98, 0.02,
-        f"$\\chi^2_\\nu$ = {reduced_chi_squared_fit:.2f}",
-        transform=plt.gca().transAxes,
-        fontsize=10,
-        verticalalignment='bottom',
-        horizontalalignment='right',
-        bbox=dict(facecolor='white', alpha=0.6, edgecolor='none')
-    )
-    save_path = get_output_path(
-        folder_path, folder_name, "source shift plot", extension=PLOT_EXTENSION, subfolder="plots"
-    )
-    plt.savefig(save_path)
-    plt.close()
-    logger.info(f"Saved: {save_path}")
-    # plt.show()
-
-
-# ---- Function for Analysis Type 4: Photon Tally Plot ----
-def run_analysis_type_4(file_path, export_csv=True):
-    file_path = Path(file_path)
-    _, df_photon = read_tally_blocks_to_df(file_path)
-    if df_photon is None or df_photon.empty:
-        logger.warning("No photon tally data found.")
-        return
-
-    # --- Save photon tally to CSV ---
-    base_name = file_path.stem
-    base_dir = file_path.parent
-    if export_csv:
-        photon_csv_path = get_output_path(base_dir, base_name, "photon tally", extension="csv", subfolder="csvs")
-        df_photon.to_csv(photon_csv_path, index=False)
-        logger.info(f"Saved: {photon_csv_path}")
-
-    plt.figure(figsize=(10, 6))
-    plt.plot(df_photon["photon_energy"], df_photon["photons"], label="Photons")
-    plt.xlabel("Photon Energy (MeV)")
-    plt.ylabel("Photon Counts")
-    plt.title("Photon Tally (Tally 34)")
-    plt.grid(True)
-    plt.legend()
-    plt.tight_layout()
-
-    save_path = get_output_path(
-        base_dir, base_name, "photon tally plot", extension=PLOT_EXTENSION, subfolder="plots"
-    )
-    plt.savefig(save_path)
-    plt.close()
-    logger.info(f"Saved: {save_path}")
-    # plt.show()
-
-
-# ---- Main execution logic ----
-def main(export_csv=True):
-    neutron_yield = select_neutron_yield()
-    while True:
-        Tk().withdraw()
-        analysis_type = input(
-            "Select analysis type:\n"
-            "1. Efficiency & Neutron Rates (single simulated CSV)\n"
-            "2. Thickness Comparison (one or more simulated datasets)\n"
-            "3. Source Position Alignment (varying source distance, no moderator)\n"
-            "4. Photon Tally Plot (single simulated CSV)\n"
-            "Enter 1, 2, 3, or 4: "
-        ).strip()
-
-        if analysis_type == "1":
-            file_path, _ = prompt_for_valid_file("Select MCNP Output File")
-            run_analysis_type_1(file_path, AREA, VOLUME, neutron_yield, export_csv)
-        elif analysis_type == "2":
-            try:
-                count = int(input("How many simulated datasets? ").strip())
-            except ValueError:
-                logger.warning("Invalid number entered.")
-                continue
-            folder_paths = []
-            labels = []
-            for i in range(count):
-                folder = select_folder(f"Select Folder for dataset {i+1}")
-                if not folder:
-                    logger.warning("No folder selected.")
-                    continue
-                label = input(f"Label for dataset {i+1} (leave blank for folder name): ").strip()
-                if not label:
-                    label = Path(folder).name
-                folder_paths.append(Path(folder))
-                labels.append(label)
-            lab_data_path = select_file("Select Experimental Lab Data CSV (Cancel to skip)")
-            if not lab_data_path:
-                lab_data_path = None
-            run_analysis_type_2(folder_paths, labels, lab_data_path, AREA, VOLUME, neutron_yield, export_csv)
-        elif analysis_type == "3":
-            folder_path = select_folder("Select Folder with Simulated Source Position CSVs")
-            if not folder_path:
-                logger.warning("No folder selected.")
-                continue
-            run_analysis_type_3(folder_path, AREA, VOLUME, neutron_yield, export_csv)
-        elif analysis_type == "4":
-            file_path, _ = prompt_for_valid_file("Select MCNP Output File for Gamma Analysis")
-            run_analysis_type_4(file_path, export_csv)
-        else:
-            logger.warning("Invalid selection.")
-
-        cont = input("\nWould you like to run another analysis? (Y/N): ").strip().lower()
-        if cont != "y":
-            logger.info("Exiting analysis tool.")
-            break
-
-
-# ---- Entry point ----
-if __name__ == "__main__":
-    main()
-
